requirements.txt

python-dotenv
pynytimes



